Commit with working feature file. Did a fix on the amount type. Changed this to double input.

+++ app/src/main/java/com/carlomatulessy/demobankapp/DataBuilder.java
@@ -54,10 +54,10 @@ public class DataBuilder {
     }
 
     public static void updateUserAmount() {
-        int userCurrentAmount = user.getAmount();
-        int inputAccountAmount = inputAccount.getAmount();
+        double userCurrentAmount = user.getAmount();
+        double inputAccountAmount = inputAccount.getAmount();
 
-        int userNewAmount = userCurrentAmount - inputAccountAmount;
+        double userNewAmount = userCurrentAmount - inputAccountAmount;
 
         user.setAmount(userNewAmount);
     }
